review(kylehickinson): shift shared logic from delegate to helper function.
